+++ changelog.md
@@ -9,10 +9,12 @@
   * Crystals
     * Cages and More Crystals now generate starting from smaller towers and no longer randomly
     * Crystals respawned by Dragon now always start from the bigger towers to the smaller ones instead of begin random. (0.2 ~~tries to respawn a crystal~~ crystals respawned per difficulty, rounded down)
-    * Changed crystal respawning health threshold (0% chance when health >= 50%; 100% chance when health <= ~~20%~~ -> 15%)
+    * Changed crystal respawning health threshold (0% chance when health >= ~~50%~~ -> 75%; 100% chance when health <= ~~20%~~ -> 15%) (health threshold reduced by ~~20%~~ -> 30% every time the crystals are respawned)
     * Time to respawn a crystal is now fixed (~~50~44 ticks based off difficulty~~ -> 50 ticks)
   * Attack
     * Dragon will no longer fireball the player if too near
+  * Resistances
+    * Increased Bonus direct damage when not in the center (~~2%~~ -> 2.5% of current health)
 
 ## 3.5.2
 * Dragon